검 공격 시 회전

+++ Assets/02_Scripts/Core/Timer/TimerHelper.cs
@@ -8,7 +8,7 @@ public static class TimerHelper
     {
 
         {
-            typeof(int),
+            typeof(float),
             new FloatTimer()
         },
